[MouseWithoutBorders] - moving Common.Service.cs -> Core\Service.cs - #35155
